Rename pkg forgotpwdhandler to forgotpwd
