+++ pkg/auth/handler/forgotpwd/forgot_password_test.go
@@ -1,4 +1,4 @@
-package forgotpwdhandler
+package forgotpwd
 
 import (
 	"net/http"
