+++ pkg/auth/handler/forgotpwd/handler.go
@@ -1,4 +1,4 @@
-package forgotpwdhandler
+package forgotpwd
 
 import "time"
