+++ pkg/auth/handler/forgotpwd/reset.go
@@ -1,4 +1,4 @@
-package forgotpwdhandler
+package forgotpwd
 
 import (
 	"encoding/json"
