+++ pkg/auth/handler/forgotpwd/reset_form.go
@@ -1,4 +1,4 @@
-package forgotpwdhandler
+package forgotpwd
 
 import (
 	"io"
